fixed the comma

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,15 +7,15 @@
  */
 int main(void)
 {
-	int res;
-	int n1 = 0;
-	int n2 = 1;
+	unsigned long int res;
+	unsigned long int n1 = 0;
+	unsigned long int n2 = 1;
 	int count;
 
 	for (count = 1; count <= 98; count++)
 	{
 		res = n1 + n2;
-		printf("%d, ",res);
+		printf("%ld, ",res);
 		n1 = n2;
 		n2 = res;
 	}
